Bugfixes for compilation.  Also added open loop benchmark to main file.

+++ core-core/print_results.h
@@ -15,5 +15,5 @@
     void print_results_blocked_fifo_standalone(Results &results, int elements_per_trial, int block_length, int bytes_per_element);
     void print_results_blocked_fifo(Results &results, int elements_per_trial, int length, int block_length, int bytes_per_element, std::string format, FILE* file=NULL, std::ofstream* raw_file=NULL);
     void print_results_open_loop_standalone(Results &results);
-    void print_results_open_loop(Results &results, size_t array_length, size_t block_length, int nops, std::string format, FILE* file, std::ofstream* raw_file);
+    void print_results_open_loop(Results &results, size_t array_length, int32_t block_length, int32_t nops, std::string format, FILE* file, std::ofstream* raw_file);
 #endif
